fix:opencode
