+++ demo/sspa/src/root-config.ts
@@ -1,5 +1,6 @@
 // demo/sspa/src/root-config.ts
 import { registerApplication, start } from 'single-spa';
+import type { LifeCycles } from 'single-spa';
 
 type DiscoveryApplication = {
   name: string;
@@ -8,7 +9,7 @@ type DiscoveryApplication = {
 };
 
 declare const System: {
-  import: (moduleName: string) => Promise<unknown>;
+  import: (moduleName: string) => Promise<LifeCycles<{}>>;
 };
 
 const DISCOVERY_BASE_URL = import.meta.env.VITE_DISCOVERY_BASE_URL || 'http://localhost:3015';
