+++ demo/sspa/src/vite-env.d.ts
@@ -1,0 +1,2 @@
+// demo/sspa/src/vite-env.d.ts
+/// <reference types="vite/client" />
