+++ gen/static-mfe/src/Bootstrap.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import ReactDOM from 'react-dom/client';
 import App from './App';
 
 const Bootstrap = () => {
