+++ gen/static-mfe/src/single-spa.ts
@@ -1,22 +1,2 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import singleSpaReact from 'single-spa-react';
-import App from './App';
-
-const lifecycles = singleSpaReact({
-  React,
-  ReactDOM,
-  rootComponent: App,
-  errorBoundary(err, props) {
-    return (
-      <div>
-        <h3>Erro no componente</h3>
-        <p>{err?.message}</p>
-      </div>
-    );
-  },
-});
-
-export const bootstrap = lifecycles.bootstrap;
-export const mount = lifecycles.mount;
-export const unmount = lifecycles.unmount;
+// gen/static-mfe/src/single-spa.ts
+export { bootstrap, mount, unmount } from './spa';
