+++ gen/static-mfe/src/types/single-spa-react.d.ts
@@ -1,0 +1,2 @@
+// gen/static-mfe/src/types/single-spa-react.d.ts
+declare module 'single-spa-react';
