+++ gen/templates/mfe-list-page.ejs
@@ -5,7 +5,6 @@ import { ErrorBoundary } from '../components/ErrorBoundary';
 
 interface <%= pascalName %>Data {
   id: <%= idType %>;
-  external_id?: string;
   created_at?: string;
   updated_at?: string;
 <% columns.forEach(function(col) { %>
