+++ gen/templates/mfe-package-json.ejs
@@ -23,6 +23,6 @@
     "@vitejs/plugin-react": "^4.2.1",
     "typescript": "^5.3.3",
     "vite": "^5.0.8",
-    "vite-plugin-single-spa": "^5.0.0"
+    "vite-plugin-single-spa": "latest"
   }
 }
